+++ src/interface/guide/components/TimelineHeatmapGrid.module.scss
@@ -9,19 +9,20 @@
   width: 100%;
 }
 
-.CornerCell {
-  background: transparent;
+.AxisBarLine {
+  align-self: start;
+  height: 2px;
+  background: rgba(255, 255, 255, 0.15);
 }
 
-.TimeAxis {
+.AxisLabels {
   position: relative;
   height: 100%;
-  border-bottom: 2px solid rgba(255, 255, 255, 0.15);
 }
 
 .TimeAxisTick {
   position: absolute;
-  bottom: 2px;
+  top: 0px;
   transform: translateX(-50%);
   font-size: 1.2rem;
   font-weight: 600;
@@ -38,31 +39,26 @@
 }
 
 .TotalHeader {
+  align-self: start;
   display: flex;
-  align-items: flex-end;
+  align-items: center;
   justify-content: center;
-  padding-bottom: 2px;
+  padding-top: 6px;
   font-size: 1.2rem;
   font-weight: 700;
   color: rgba(255, 255, 255, 0.85);
   white-space: nowrap;
-  border-bottom: 2px solid rgba(255, 255, 255, 0.25);
+  border-top: 2px solid rgba(255, 255, 255, 0.15);
 }
 
 .RowLabel {
   display: flex;
   align-items: center;
+  justify-content: flex-end;
   font-size: 1.4rem;
   font-weight: 700;
-  padding-left: 10px;
-  border-left: 4px solid transparent;
-}
-
-.GradientTrack {
-  align-self: center;
-  height: 12px;
-  border-radius: 6px;
-  box-shadow: inset 0 0 0 1px rgba(255, 255, 255, 0.12);
+  padding-right: 10px;
+  border-right: 4px solid transparent;
 }
 
 .Cell {
@@ -82,12 +78,12 @@
 .TotalCell {
   display: flex;
   flex-direction: column;
-  align-items: center;
+  align-items: flex-start;
   justify-content: center;
   gap: 2px;
   line-height: 1.1;
-  border-right: 4px solid transparent;
-  padding: 0 10px 0 0;
+  border-left: 4px solid transparent;
+  padding: 0 0 0 10px;
 }
 
 .TotalPercent {
